+++ R/broom.R
@@ -30,7 +30,12 @@ augment.mdl_df <- function(x, ...){
 
 #' @rdname augment
 #' @export
-augment.mdl_ts <- function(x, ...){
+augment.mdl_ts <- function(x, type = NULL, ...){
+  if (!is.null(type)) { 
+    lifecycle::deprecate_warn("0.2.1", "fabletools::augment(type = )", 
+                              details = "The type argument is now deprecated for changes to broom v0.7.0.
+Response residuals are now always found in `.resid` and innovation residuals are now found in `.innov`.")
+  }
   tryCatch(augment(x[["fit"]], ...),
            error = function(e){
              idx <- index_var(x$data)
@@ -44,7 +49,12 @@ augment.mdl_ts <- function(x, ...){
                    by = c(".response", idx)
                  ) %>% 
                  left_join(
-                   gather(residuals(x, ...), ".response", ".resid",
+                   gather(residuals(x, type = "response", ...), ".response", ".resid",
+                          !!!resp, factor_key = TRUE),
+                   by = c(".response", idx)
+                 ) %>% 
+                 left_join(
+                   gather(residuals(x, type = "innovation", ...), ".response", ".innov",
                           !!!resp, factor_key = TRUE),
                    by = c(".response", idx)
                  )
@@ -52,7 +62,8 @@ augment.mdl_ts <- function(x, ...){
                mutate(
                  set_names(response(x), c(idx, as_string(resp[[1]]))),
                  .fitted = fitted(x, ...)[[".fitted"]],
-                 .resid = residuals(x, ...)[[".resid"]]
+                 .resid = residuals(x, type = "response", ...)[[".resid"]],
+                 .innov = residuals(x, type = "innovation", ...)[[".resid"]],
                )
              }